Move BlockEvent from events-world to extensions-block.
This removes a unnecessary dependency.
