+++ patchwork-events-world/build.gradle
@@ -3,5 +3,4 @@ version = getSubprojectVersion(project, "0.3.0")
 
 dependencies {
 	implementation project(path: ':patchwork-api-base', configuration: 'dev')
-	implementation project(path: ':patchwork-extensions-block', configuration: 'dev')
 }
