+++ patchwork-events-world/src/main/java/net/patchworkmc/impl/event/world/WorldEvents.java
@@ -23,24 +23,13 @@ import java.util.Collections;
 import java.util.List;
 
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.event.world.ChunkWatchEvent;
 import net.minecraftforge.event.world.WorldEvent;
 
-import net.minecraft.block.BlockState;
-import net.minecraft.block.entity.BlockEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.DefaultedList;
-import net.minecraft.world.World;
 import net.minecraft.entity.EntityCategory;
-import net.minecraft.network.packet.s2c.play.BlockEntityUpdateS2CPacket;
-import net.minecraft.network.packet.s2c.play.BlockUpdateS2CPacket;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.EmptyBlockView;
-import net.minecraft.world.GameMode;
 import net.minecraft.util.math.ChunkPos;
 import net.minecraft.world.IWorld;
 import net.minecraft.world.biome.Biome;
@@ -75,67 +64,6 @@ public class WorldEvents implements ModInitializer {
 
 	public static void onWorldSave(IWorld world) {
 		MinecraftForge.EVENT_BUS.post(new WorldEvent.Save(world));
-	}
-
-	/**
-	 * Called by Mixin and ForgeHooks.
-	 * @return experience dropped, -1 = block breaking is cancelled.
-	 */
-	public static int onBlockBreakEvent(World world, GameMode gameMode, ServerPlayerEntity player, BlockPos pos) {
-		// Logic from tryHarvestBlock for pre-canceling the event
-		boolean preCancelEvent = false;
-
-		ItemStack itemstack = player.getMainHandStack();
-
-		if (!itemstack.isEmpty() && !itemstack.getItem().canMine(world.getBlockState(pos), world, pos, player)) {
-			preCancelEvent = true;
-		}
-
-		// method_21701 => canMine
-		// Isn't the function really canNotMine?
-
-		if (player.method_21701(world, pos, gameMode)) {
-			preCancelEvent = true;
-		}
-
-		// Tell client the block is gone immediately then process events
-		if (world.getBlockEntity(pos) == null) {
-			player.networkHandler.sendPacket(new BlockUpdateS2CPacket(EmptyBlockView.INSTANCE, pos));
-		}
-
-		// Post the block break event
-		BlockState state = world.getBlockState(pos);
-		BlockEvent.BreakEvent event = new BlockEvent.BreakEvent(world, pos, state, player);
-		event.setCanceled(preCancelEvent);
-		MinecraftForge.EVENT_BUS.post(event);
-
-		// Handle if the event is canceled
-		if (event.isCanceled()) {
-			// Let the client know the block still exists
-			player.networkHandler.sendPacket(new BlockUpdateS2CPacket(world, pos));
-
-			// Update any block entity data for this block
-			BlockEntity entity = world.getBlockEntity(pos);
-
-			if (entity != null) {
-				BlockEntityUpdateS2CPacket packet = entity.toUpdatePacket();
-
-				if (packet != null) {
-					player.networkHandler.sendPacket(packet);
-				}
-			}
-
-			return -1; // Cancelled
-		} else {
-			return event.getExpToDrop();
-		}
-	}
-
-	// TODO: Leaving this unfired is intentional. See: https://github.com/MinecraftForge/MinecraftForge/issues/5828
-	public static float fireBlockHarvesting(DefaultedList<ItemStack> drops, World world, BlockPos pos, BlockState state, int fortune, float dropChance, boolean silkTouch, PlayerEntity player) {
-		BlockEvent.HarvestDropsEvent event = new BlockEvent.HarvestDropsEvent(world, pos, state, fortune, dropChance, drops, player, silkTouch);
-		MinecraftForge.EVENT_BUS.post(event);
-		return event.getDropChance();
 	}
 
 	public static void fireChunkWatch(boolean watch, ServerPlayerEntity entity, ChunkPos chunkpos, ServerWorld world) {
