+++ patchwork-events-world/src/main/resources/fabric.mod.json
@@ -13,8 +13,7 @@
     "PatchworkMC"
   ],
   "depends": {
-    "patchwork-api-base": "*",
-    "patchwork-extensions-block": "*"
+    "patchwork-api-base": "*"
   },
   "mixins": [
     "patchwork-events-world.mixins.json"
