+++ patchwork-events-world/src/main/resources/patchwork-events-world.mixins.json
@@ -5,10 +5,9 @@
   "mixins": [
     "MixinChunkGenerator",
     "MixinMinecraftServer",
-    "MixinServerPlayerInteractionManager",
     "MixinServerWorld",
     "MixinSpawnHelper",
-    "MixinThreadedAnvilChunkStorage",
+    "MixinThreadedAnvilChunkStorage"
   ],
   "client": [
     "MixinClientWorld",
