+++ patchwork-god-classes/src/main/resources/fabric.mod.json
@@ -18,9 +18,12 @@
     "patchwork-api-base": "*",
     "patchwork-fml": "*",
     "patchwork-capabilities": "*",
+    "patchwork-events-entity": "*",
     "patchwork-events-input": "*",
     "patchwork-events-lifecycle": "*",
     "patchwork-events-rendering": "*",
+    "patchwork-events-world": "*",
+    "patchwork-extensions-block": "*",
     "patchwork-loot": "*"
   },
   "custom": {
